+++ source/c_bitmap.cpp
@@ -689,7 +689,7 @@ HB_FUNC( DRAWGLYPHMASK )
    // handle to bitmaped button mask
    if( hwnd != nullptr )
    {
-      SendMessage(hwnd, ( UINT ) BM_SETIMAGE, IMAGE_BITMAP, ( LPARAM ) hBmpTransMask);
+      SendMessage(hwnd, BM_SETIMAGE, IMAGE_BITMAP, ( LPARAM ) hBmpTransMask);
    }
 
    SelectObject(hDCMem, hBmpDefault);
