+++ source/c_getbox.cpp
@@ -166,7 +166,7 @@ HB_FUNC( INITGETBOX )
    SetProp(hedit, TEXT("OldWndProc"), reinterpret_cast<HWND>(GetWindowLongPtr(hedit, GWLP_WNDPROC)));
    SetWindowLongPtr(hedit, GWLP_WNDPROC, ( LONG_PTR ) ( WNDPROC ) OwnGetProc);
 
-   SendMessage(hedit, ( UINT ) EM_LIMITTEXT, hmg_par_WPARAM(9), 0);
+   SendMessage(hedit, EM_LIMITTEXT, hmg_par_WPARAM(9), 0);
 
    if( hb_parc(18) != nullptr )
    {
@@ -305,12 +305,12 @@ HB_FUNC( INITGETBOX )
 
    if( himage != nullptr )
    {
-      SendMessage(hBtn1, ( UINT ) BM_SETIMAGE, IMAGE_BITMAP, ( LPARAM ) himage);
+      SendMessage(hBtn1, BM_SETIMAGE, IMAGE_BITMAP, ( LPARAM ) himage);
    }
 
    if( himage2 != nullptr )
    {
-      SendMessage(hBtn2, ( UINT ) BM_SETIMAGE, IMAGE_BITMAP, ( LPARAM ) himage2);
+      SendMessage(hBtn2, BM_SETIMAGE, IMAGE_BITMAP, ( LPARAM ) himage2);
    }
 
    SendMessage(hedit, EM_SETMARGINS, EC_LEFTMARGIN | EC_RIGHTMARGIN, MAKELONG(0, BtnWidth + BtnWidth2 + 2));
