+++ source/c_textbox.cpp
@@ -193,7 +193,7 @@ HB_FUNC( INITTEXTBOX )
       nullptr
            );
 
-   SendMessage(hedit, ( UINT ) EM_LIMITTEXT, hmg_par_WPARAM(9), 0);
+   SendMessage(hedit, EM_LIMITTEXT, hmg_par_WPARAM(9), 0);
 
    SetProp(hedit, TEXT("oldeditproc"), reinterpret_cast<HWND>(GetWindowLongPtr(hedit, GWLP_WNDPROC)));
    SetWindowLongPtr(hedit, GWLP_WNDPROC, ( LONG_PTR ) ( WNDPROC ) OwnEditProc);
